Add schema trait and async process docs
