--- src/Queue/ClientJobSchema.php
@@ -1,65 +1,0 @@
-<?php
-/**
- * Copyright 2018 Cloud Creativity Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace CloudCreativity\LaravelJsonApi\Queue;
-
-use Neomerx\JsonApi\Schema\SchemaProvider;
-
-abstract class ClientJobSchema extends SchemaProvider
-{
-
-    /**
-     * @var string|null
-     */
-    protected $api;
-
-    /**
-     * @return string
-     */
-    public function getResourceType()
-    {
-        return json_api($this->api)->getJobs()->getResource();
-    }
-
-    /**
-     * @param ClientJob $resource
-     * @return string
-     */
-    public function getId($resource)
-    {
-        return (string) $resource->getRouteKey();
-    }
-
-    /**
-     * @param ClientJob|null $resource
-     * @return string
-     */
-    public function getSelfSubUrl($resource = null)
-    {
-        if (!$resource) {
-            return parent::getSelfSubUrl();
-        }
-
-        return sprintf(
-            '/%s/%s/%s',
-            $resource->resource_type,
-            $this->getResourceType(),
-            $this->getId($resource)
-        );
-    }
-
-}
